backend/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
django-environ>=0.4.5
django-extensions==3.1.3
django-stubs==1.8.0
djangorestframework
django-cors-headers
PyJWT==1.7.1
djangorestframework-jwt==1.11.0
argon2-cffi
pyjwt
pytz
requests